cleanup | more meaningful filenames, move test-files to better positions
