+++ api/middlewares/__test__/errorHandler.test.ts
@@ -1,8 +1,4 @@
-import { Request, Response } from 'express';
-import errorHandler from '../utils/errorHandler';
 import supertest from 'supertest';
-import app from '../utils/express';
-import logger from '../utils/logger';
 
 describe('Error Handler', () => {
   beforeEach(() => {
@@ -13,7 +9,7 @@ describe('Error Handler', () => {
   it('should include stack in the response in development mode', async () => {
     process.env.NODE_ENV = 'development';
 
-    const { default: app } = require('../utils/express');
+    const { default: app } = require('../../utils/express');
 
     const path = '/nope';
     const { statusCode, body } = await supertest(app).get(path);
@@ -25,7 +21,7 @@ describe('Error Handler', () => {
 
   it('should not include stack in the response in production mode', async () => {
     process.env.NODE_ENV = 'production';
-    const { default: app } = require('../utils/express');
+    const { default: app } = require('../../utils/express');
 
     const path = '/nope';
     const { statusCode, body } = await supertest(app).get(path);
@@ -37,7 +33,7 @@ describe('Error Handler', () => {
 
   it('should not include stack in the response in test mode', async () => {
     process.env.NODE_ENV = 'test';
-    const { default: app } = require('../utils/express');
+    const { default: app } = require('../../utils/express');
 
     const path = '/nope';
     const { statusCode, body } = await supertest(app).get(path);
