+++ api/routes/__test__/index.test.ts
@@ -1,5 +1,5 @@
 import supertest from 'supertest';
-import app from '../utils/express';
+import app from '../../utils/express';
 
 describe('default', () => {
   // ----------------------------------------------------------------
