+++ api/utils/express.ts
@@ -2,8 +2,8 @@ import bodyParser from 'body-parser';
 import cors from 'cors';
 import express from 'express';
 import routes from '../routes';
-import corsOptions from './corsOptions';
-import errorHandler from './errorHandler';
+import corsOptions from '../config/cors.config';
+import errorHandler from '../middlewares/errorHandler';
 import swaggerDocs from './swagger';
 import createHttpError from 'http-errors';
 import 'express-async-errors';
