+++ api/utils/swagger.ts
@@ -3,7 +3,7 @@ import swaggerJSDoc from 'swagger-jsdoc';
 import swaggerUi from 'swagger-ui-express';
 import { version } from '../../package.json';
 import logger from './logger';
-import { PORT, SERVER } from './environment';
+import { PORT, SERVER } from '../config/environment.config';
 
 const options: swaggerJSDoc.Options = {
   definition: {
