+++ api/validators/order.validator.ts
@@ -1,6 +1,6 @@
 import { body, param } from 'express-validator';
 import { isValidObjectId } from 'mongoose';
-import { PizzaStatus } from '../models/order';
+import { PizzaStatus } from '../models/order.model';
 
 export const createOrder = [
   body('customer.name').notEmpty().isString(),
